removed Wiki functionality and refactored
