+++ database.go
@@ -1,88 +1,6 @@
 package main
 
 import (
-	"database/sql"
-	"fmt"
-	"net/http"
-
+	// Comment
 	_ "github.com/go-sql-driver/mysql"
-	"golang.org/x/crypto/bcrypt"
 )
-
-//GLOBAL VARIABLES
-var db *sql.DB
-var err error
-
-func loginPage(res http.ResponseWriter, req *http.Request) {
-	// If method is GET serve an html login page
-	if req.Method != "POST" {
-		http.ServeFile(res, req, "login.html")
-		return
-	}
-	// Grab the username/password from the submitted post form
-	username := req.FormValue("username")
-	password := req.FormValue("password")
-
-	// Grab from the database
-	var databaseUsername string
-	var databasePassword string
-
-	// Search the database for the username provided
-	// If it exists grab the password for validation
-	err := db.QueryRow("SELECT username, password FROM users WHERE username=?", username).Scan(&databaseUsername, &databasePassword)
-	// If not then redirect to the login page
-	if err != nil {
-		http.Redirect(res, req, "/login", 301)
-		return
-	}
-
-	// Validate the password
-	err = bcrypt.CompareHashAndPassword([]byte(databasePassword), []byte(password))
-	// If wrong password redirect to the login
-	if err != nil {
-		http.Redirect(res, req, "/login", 301)
-		return
-	}
-
-	// If the login succeeded
-	res.Write([]byte("Hello " + databaseUsername))
-}
-
-func signupPage(res http.ResponseWriter, req *http.Request) {
-
-	// Serve signup.html to get requests to /signup
-	if req.Method != "POST" {
-		http.ServeFile(res, req, "signup.html")
-		return
-	}
-
-	username := req.FormValue("username")
-	password := req.FormValue("password")
-	fmt.Printf("username: %s Password: %s", username, password)
-	//var user string
-
-	//err := db.QueryRow("SELECT username FROM users WHERE username=?", username).Scan(&user)
-	switch {
-	// Username is available
-	case err == sql.ErrNoRows:
-		hashedPassword, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-		if err != nil {
-			http.Error(res, "Server error, unable to create your account.", 500)
-			return
-		}
-
-		_, err = db.Exec("INSERT INTO users(username, password) VALUES(?, ?)", username, hashedPassword)
-		if err != nil {
-			http.Error(res, "Server error, unable to create your account.", 500)
-			return
-		}
-
-		res.Write([]byte("User created!"))
-		return
-	case err != nil:
-		http.Error(res, "Server error, unable to create your account.", 500)
-		return
-	default:
-		http.Redirect(res, req, "/", 301)
-	}
-}
