+++ sample.txt
@@ -3,3 +3,4 @@ Hello World <\r>
 Good Evening <\n>
 
 Another Day 
+
